Move docker logic into a separate file
Since docker setup is not related to minikube anymore
